implement combinator and complex selector matching
